Fixed While loop
Was relying on undefined returned value to stop, but this worked because
of another bug: Blocks were reducing stats to an array of promises,
which was never undefined, but this array could contain undefined.

But reducing should have returned a promise all the time. When the bug
was fixed, it revealed the other bug.